Minor fixes from sync comment.
* Typo in comment
* Round timeout float before passing to Java.

+++ mobly/controllers/android_device_lib/callback_handler.py
@@ -85,7 +85,8 @@ class CallbackHandler(object):
                 raise Error(
                     'Specified timeout %s is longer than max timeout %s.' %
                     (timeout, MAX_TIMEOUT))
-        timeout_ms = timeout * 1000  # convert to milliseconds for java side
+        timeout_ms = int(
+            timeout * 1000)  # convert to milliseconds for java side
         try:
             raw_event = self._event_client.eventWaitAndGet(
                 self._id, event_name, timeout_ms)
